% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_data_summary.R
\docType{data}
\name{station_data_summary}
\alias{station_data_summary}
\title{INMET stations data}
\format{Data frame with columns
\describe{
\item{station_id}{Station ID.}
\item{date}{Record date.}
\item{prec}{Total rainfall in mm.}
\item{temp_max}{Maximum temperature in degrees celcius.}
\item{temp_min}{Minimum temperature in degrees celcius.}
\item{temp_mean}{Mean temperature in degrees celcius.}
\item{insol}{Insolation hours.}
\item{humid}{Mean relative humidity as percentage.}
}}
\source{
INMET, Instituto Nacional de Meteorologia, \url{http://inmet.gov.br}
}
\usage{
station_data_summary
}
\description{
Climate dataset compiled from information available from INMET for 261
climate stations located accross Brazil. Dates range from 1961-2017 depending
on the station.
}
\keyword{datasets}
