% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_info.R
\docType{data}
\name{station_info}
\alias{station_info}
\title{INMET stations information}
\format{Data frame with columns
\describe{
\item{station_id}{Station ID.}
\item{start}{Start of the available period.}
\item{end}{End of the available period.}
\item{complete}{Logical indicating if there are days with missing data.}
\item{days_missing}{Number of days missing.}
\item{locality}{Station locality.}
\item{state}{Station state.}
\item{latitude and longitude}{Station coordinates.}
}}
\source{
INMET, Instituto Nacional de Meteorologia, \url{http://inmet.gov.br}
}
\usage{
station_info
}
\description{
Location and data range and completeness of 261 INMET stations located in Brazil.
}
\keyword{datasets}
