% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_data_summary.R
\docType{data}
\name{station_data_summary}
\alias{station_data_summary}
\title{INMET stations data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3952876 rows and 8 columns.}
\source{
INMET, Instituto Nacional de Meteorologia, \url{http://inmet.gov.br}
}
\usage{
station_data_summary
}
\description{
Climate dataset compiled from information available from INMET for 261
climate stations located accross Brazil. Dates range from 1961-2017 depending
on the station.
}
\keyword{datasets}
